requirements.txt

PyPDF2 
pdfplumber
openai
pandas
python-dotenv
